+++ tensorflow_quantum/core/ops/tfq_ps_symbol_replace_op.cc
@@ -163,11 +163,11 @@ class TfqPsSymbolReplaceOp : public tensorflow::OpKernel {
       for (int i = start; i < end; i++) {
         int sidx = i % n_symbols;
         int pidx = i / n_symbols;
-        for (size_t j = 0; j < output_programs.at(pidx).at(sidx).size(); j++) {
+        for (int j = 0; j < output_programs.at(pidx).at(sidx).size(); j++) {
           output_tensor(pidx, sidx, j) =
               output_programs.at(pidx).at(sidx).at(j);
         }
-        for (size_t j = output_programs.at(pidx).at(sidx).size(); j < biggest_pad;
+        for (int j = output_programs.at(pidx).at(sidx).size(); j < biggest_pad;
              j++) {
           output_tensor(pidx, sidx, j) = empty_program;
         }
